pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hf-ai-agent-course-capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.28.0",
    "langchain>=0.3.24",
    "langchain-community>=0.3.23",
    "langchain-google-genai>=2.1.3",
    "langchain-mistralai>=0.2.10",
    "langchain-openai>=0.3.14",
    "litellm==1.66.3",
    "llama-index-readers-youtube-transcript>=0.3.0",
    "openai-whisper>=20240930",
    "openpyxl>=3.1.5",
    "pydub>=0.25.1",
    "python-ffmpeg>=2.0.12",
    "pytubefix>=8.13.1",
    "smolagents[litellm]==1.9.2",
    "soundfile>=0.13.1",
    "speechrecognition>=3.14.2",
    "transformers>=4.51.3",
    "vosk>=0.3.45",
    "wikipedia>=1.4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "hatch>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/azrock", "tests"]
